Refactored the entire FileBinaryReader.java
